fix: break things into components and use events
