+++ packages/bruno-app/src/components/Notifications/NotificationsModal/NotificationList.js
@@ -1,0 +1,25 @@
+import { relativeDate } from 'utils/common';
+
+const NotificationList = ({ items, selectedId, onSelect }) => {
+  return (
+    <ul className="notif-list">
+      {items.map((notification) => {
+        const isActive = selectedId === notification.id;
+        const isUnread = !notification.read;
+        return (
+          <li
+            key={notification.id}
+            className={`notif-list-item ${isActive ? 'active' : ''} ${isUnread ? 'unread' : ''}`}
+            onClick={() => onSelect(notification)}
+          >
+            <div className={`notif-item-title ${isUnread ? 'unread' : ''}`}>{notification.title}</div>
+            <div className="notif-item-date">{relativeDate(notification.date)}</div>
+          </li>
+        );
+      })}
+      {items.length === 0 && <li className="notif-list-empty">No notifications to show.</li>}
+    </ul>
+  );
+};
+
+export default NotificationList;
